Add files via upload
This is a terrible and probably wrong finite difference solver for the GS equation. There is version is built into the main class, and another version that is a standalone function. The standalone one needs the code that I added to the main class to output the flux to a file, as it reads flux from that file. This is all ugly, but until I can get the bloody thing to work properly I didn't see much point in making it pretty/vectorised.