Add files via upload
This is an almost complete rebuild of our main class. It has the numerical greens, so it should work with numba, but numba keeps fighting me. I also wasn't satisfied with how the miller surface was implemented previously, so I am trying to make a better way of doing it. That's still incoming. For now, it just runs symmetric solovev. I haven't had a chance to really push the grid size as pre numba it is slow, but it seems to largely work. My only issue is that I wanted LCFS matching to be done with points that actually have plasma in them rather than potential algebraic ones, so the matching is a bit janky. I am still working on that. This is just going up now so that Joe has a chance to look at the numba at some point.