Add files via upload
Added the finite difference code that can run based on the output of the main class. It also plots the current grid that the attached gs_solver outputs. The gs solver attached here fixes an issue with the current that I found in the main class (missing a factor of r_p). It also adds the current output function and implements a more physical current for the miller solution.